+++ Projectile.h
@@ -19,7 +19,7 @@ public:
     int damages() { return p_damages; }
     bool playerProjectile() { return p_players; }
     int getDirection() { return p_direction; }
-    double hitBox() { return p_hitbox; }
+    int hitBox() { return p_hitbox; }
     void move();
     bool murred() { return p_position.murred(p_direction); }
 };
